LayoutTests/imported/w3c/web-platform-tests/dom/events/Event-timestamp-cross-realm-getter-expected.txt



PASS event.timeStamp is initialized using event's relevant global object
